move time override outside admin
