+++ app/javascript/stylesheets/application.scss
@@ -1098,3 +1098,9 @@ footer.page-footer {
   text-align: center;
   position: relative;
 }
+
+.time-override {
+  position: absolute;
+  top: 0;
+  right: 0;
+}
