+++ app/views/shared/_time_override.html.erb
@@ -1,12 +1,12 @@
-<section class="navbar-section">
+<div class="time-override">
   <div class="dropdown">
-    <a href="#" class="btn btn-link dropdown-toggle" tabindex="0" data-turbolinks="false">
+    <a href="#" class="btn btn-error btn-sm dropdown-toggle" tabindex="0" data-turbolinks="false">
       Change Time <i class="icon icon-caret"></i>
     </a>
     <ul class="menu">
       <li class="divider" data-content="Set time">
       <li class="menu-item">
-        <%= form_with url: admin_dev_time_path, class: "form" do |form| %>
+        <%= form_with url: dev_time_path, class: "form" do |form| %>
           <div class="input-group input-inline">
             <%= form.text_field :new_time, value: Time.zone.now.to_s, class: "form-input" %>
             <%= form.hidden_field :return_to, value: request.path %>
@@ -15,11 +15,11 @@
       </li>
       <li class="divider" data-content="Clear override">
       <li class="menu-item">
-        <%= form_with url: admin_dev_time_path, method: :delete, class: "form" do |form| %>
+        <%= form_with url: dev_time_path, method: :delete, class: "form" do |form| %>
           <%= form.hidden_field :return_to, value: request.path %>
           <%= form.button "Reset", type: "submit", class: "btn btn-sm" %>
         <% end %>
       </li>
     </ul>
   </div>
-</section>
+</div>
