+++ config/routes.rb
@@ -125,13 +125,13 @@ Rails.application.routes.draw do
     get "/ui/sizes", to: "ui#sizes"
     get "/ui/strengths", to: "ui#strengths"
 
-    if Rails.env.development?
-      post "/dev/time", to: "dev#set_time"
-      delete "/dev/time", to: "dev#clear_time"
-      get "/dev/styles", to: "dev#styles"
-    end
-
     get "/", to: "dashboard#index", as: "dashboard"
+  end
+
+  if Rails.env.development?
+    post "/dev/time", to: "dev#set_time"
+    delete "/dev/time", to: "dev#clear_time"
+    get "/dev/styles", to: "dev#styles"
   end
 
   get "/s/:id", to: "short_links#show", as: :short_link
